prepare boolean for the test, and structure it as it is growing
